+++ matrix-sdk/src/main/java/org/matrix/androidsdk/rest/client/LoginRestClient.java
@@ -18,6 +18,7 @@
 package org.matrix.androidsdk.rest.client;
 
 import android.os.Build;
+import android.support.annotation.Nullable;
 import android.text.TextUtils;
 
 import com.google.gson.JsonObject;
@@ -136,7 +137,7 @@ public class LoginRestClient extends RestClient<LoginApi> {
      * @param callback the callback success and failure callback
      */
     public void loginWithUser(final String user, final String password, final ApiCallback<Credentials> callback) {
-        loginWithUser(user, password, null, callback);
+        loginWithUser(user, password, null, null, callback);
     }
 
     /**
@@ -145,14 +146,20 @@ public class LoginRestClient extends RestClient<LoginApi> {
      * @param user       the username
      * @param password   the password
      * @param deviceName the device name
+     * @param deviceId   the device id
      * @param callback   the callback success and failure callback
      */
-    public void loginWithUser(final String user, final String password, final String deviceName, final ApiCallback<Credentials> callback) {
+    public void loginWithUser(final String user,
+                              final String password,
+                              final String deviceName,
+                              @Nullable final String deviceId,
+                              final ApiCallback<Credentials> callback) {
         final String description = "loginWithUser : " + user;
 
         PasswordLoginParams params = new PasswordLoginParams();
         params.setUserIdentifier(user, password);
         params.setDeviceName(deviceName);
+        params.setDeviceId(deviceId);
 
         login(params, callback, description);
     }
@@ -166,7 +173,7 @@ public class LoginRestClient extends RestClient<LoginApi> {
      * @param callback the callback success and failure callback
      */
     public void loginWith3Pid(final String medium, final String address, final String password, final ApiCallback<Credentials> callback) {
-        loginWith3Pid(medium, address, password, null, callback);
+        loginWith3Pid(medium, address, password, null, null, callback);
     }
 
     /**
@@ -182,12 +189,14 @@ public class LoginRestClient extends RestClient<LoginApi> {
                               final String address,
                               final String password,
                               final String deviceName,
+                              @Nullable final String deviceId,
                               final ApiCallback<Credentials> callback) {
         final String description = "loginWith3pid : " + address;
 
         PasswordLoginParams params = new PasswordLoginParams();
         params.setThirdPartyIdentifier(medium, address, password);
         params.setDeviceName(deviceName);
+        params.setDeviceId(deviceId);
 
         login(params, callback, description);
     }
@@ -201,7 +210,7 @@ public class LoginRestClient extends RestClient<LoginApi> {
      * @param callback    the callback success and failure callback
      */
     public void loginWithPhoneNumber(final String phoneNumber, final String countryCode, final String password, final ApiCallback<Credentials> callback) {
-        loginWithPhoneNumber(phoneNumber, countryCode, password, null, callback);
+        loginWithPhoneNumber(phoneNumber, countryCode, password, null, null, callback);
     }
 
     /**
@@ -211,18 +220,21 @@ public class LoginRestClient extends RestClient<LoginApi> {
      * @param countryCode the ISO country code
      * @param password    the password
      * @param deviceName  the device name
+     * @param deviceId    the device Id
      * @param callback    the callback success and failure callback
      */
     public void loginWithPhoneNumber(final String phoneNumber,
                                      final String countryCode,
                                      final String password,
                                      final String deviceName,
+                                     @Nullable final String deviceId,
                                      final ApiCallback<Credentials> callback) {
         final String description = "loginWithPhoneNumber : " + phoneNumber;
 
         PasswordLoginParams params = new PasswordLoginParams();
         params.setPhoneIdentifier(phoneNumber, countryCode, password);
         params.setDeviceName(deviceName);
+        params.setDeviceId(deviceId);
 
         login(params, callback, description);
     }